+++ RELEASE-NOTES.md
@@ -2,6 +2,7 @@
 
 * #56 Support JAX-RS Application
 * #57 bootique-jersey-jackson - add extender method to register custom type serializers
+* #58 Overlapping redundant JAXB and activation dependencies
 
 ## 2.0.M1